backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
google-generativeai==0.3.2
python-dateutil==2.8.2
pinecone-client==2.2.4
faiss-cpu>=1.7.4
numpy==1.24.3
pandas==2.1.3
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
redis==5.0.1
celery==5.3.4
httpx==0.25.2
aiofiles==23.2.1

